## Refactiorization of sprite animations. Initial commit.
### New classes.
`SpriteAnimation` class has beed introduced. It's purpose is to provide 
some generalization to the sprite animations concept, that would allow 
some more flexibility than the default `libgdx` `Animation` class. It 
takes a `TextureRegion` and internally prepares `Animation` instance.

`AnimationWalk` extends `SpriteAnimation` and provides an update method, 
that adjusts position of the sprite to be drawn to emulate movement.

For more information please read documentation.

### `Person`:
An example of `AnimationWalk` is introduced in the `draw_walk` method.

### General changes.

Changed units of variables related to screen position from `int` to 
`float` as it's the natural unit in libgdx.