## Improved animation related classes.
### `SpriteAnimation`:
Now has capabilities to change animation sprites and duration of
animation. This allows more versatile animation design and capabilities.
It also allows to recycle `SpriteAnimation` object. Because libgdx
`Animation` cannot change these propertioes on the fly. Appropriate
checks were implemented.

Also added empty constructor for constructing an unitialized objects.
With methods that allow to set up crucial parts of the animation this
makes it a little bit easier to implement extensions.

Added possibilty of calling a Finishable callback object.

### `AnimationWalk`:
Added possibilty to set direction when animation is not running.

### `PersonAnimationWalk`:
Implemented a final class for animation of walking for the `Person`.

### `Person`:
Changed walking animation to make use of new API.

### `Strategix`:
Dropped unnecessary textrure loading.